backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.3.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
av==15.0.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.12
confection==0.1.5
contourpy==1.2.1
coqpit-config==0.2.1
coqui-tts==0.27.0
coqui-tts-trainer==0.3.1
ctranslate2==4.6.0
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.8
Cython==3.1.3
dataclasses-json==0.6.7
dateparser==1.1.8
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.6
django-razorpay==1.1.6
django-storages==1.14.4
docopt==0.6.2
durationpy==0.9
einops==0.8.1
encodec==0.1.1
executing==2.0.1
Faker==26.0.0
fastapi==0.115.11
faster-whisper==1.2.0
fastjsonschema==2.19.1
filelock==3.15.4
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.5.0
fsspec==2024.6.1
funcy==2.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
googletrans==4.0.0rc1
greenlet==3.1.1
groq==0.19.0
grpcio==1.71.0
grpcio-status==1.71.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gunicorn==23.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==5.2.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.5.0
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
json5==0.9.25
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-groq==0.2.5
langchain-text-splitters==0.3.6
langcodes==3.4.0
langsmith==0.3.13
language_data==1.2.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.3.1
marisa-trie==1.2.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.1.0
monotonic==1.6
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
num2words==0.5.14
numba==0.61.2
numexpr==2.10.1
numpy==2.2.3
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.102.0
openai-whisper==20250625
opencv-python==4.10.0.84
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.6
pillow==11.1.0
platformdirs==4.2.2
pooch==1.8.2
posthog==3.20.0
preshed==3.0.9
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyLDAvis==3.4.1
pyparsing==3.2.1
PyPDF2==3.0.1
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
python-pptx==1.0.2
pyttsx3==2.99
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
razorpay==1.4.2
referencing==0.35.1
regex==2024.5.15
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.0
scipy==1.13.1
Send2Trash==1.8.3
sentence-transformers==5.1.0
sentencepiece==0.2.0
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.5
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
sqlparse==0.5.3
srsly==2.4.8
stack-data==0.6.3
starlette==0.46.1
streamlit==1.35.0
sympy==1.13.3
tenacity==8.4.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
terminado==0.18.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.11.0
tinycss2==1.3.0
tokenizers==0.21.4
toml==0.10.2
toolz==0.12.1
torch==2.8.0
torchaudio==2.8.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
typeguard==4.4.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.1
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchdog==4.0.1
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.11
wordcloud==1.9.3
wrapt==1.16.0
XlsxWriter==3.2.2
yarl==1.18.3
yfinance==0.2.40
zipp==3.21.0
zstandard==0.23.0
